Add files via upload
This is my ML Project which can predict the price of new house based on attributes .It is a Supervised learning problem regarding some features and levels .By this project ,we have done Regression task and scanned the data and visualized the Histogram models,Scattering plots. After doing that much we have seperated the dataset into two parts:
1.Transform the data(train the dataset with the models)
2. Test the model(to predict the house price as result)
Finally we have observed that the predicted result and the actual data are so close to each other & the errors are so minute.